backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.0.3
openpyxl==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.0

# Optional: For LLM-enhanced analysis (install if using actual LLM API)
# anthropic>=0.18.0
# openai>=1.0.0
